spi: read all bytes from sx12xx in one read operation
Instead of reading a single byte at a time - read all bytes with a single SPI transfer.
This is much more performant, especially when reading longer packets and when DMA is used.

Also remove the special case of read() (read_with_status()) since it's the same operation as a regular read, just with the first byte read being status.